refactor vite integration
